ds1 editor: - moved all stuff responsible for ds1 tiles to tile_records.go - moved recreateLayerStreamTypes to helper.go
